Consolidate search term management in useSearchBox
